[Refactor] 패키지 위치 이동 및 chat_message_base 연관 엔티티 양방향매핑
