+++ src/main/java/com/example/grapefield/chat/model/entity/ChatroomMember.java
@@ -1,4 +1,4 @@
-package com.example.grapefield.events.chat.model.entity;
+package com.example.grapefield.chat.model.entity;
 
 import com.example.grapefield.user.model.entity.User;
 import jakarta.persistence.*;
