+++ src/main/java/com/example/grapefield/chat/model/request/ChatMessageKafkaReq.java
@@ -1,4 +1,4 @@
-package com.example.grapefield.events.chat.model.request;
+package com.example.grapefield.chat.model.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
