+++ src/main/java/com/example/grapefield/chat/model/request/ChatMessageReq.java
@@ -1,4 +1,4 @@
-package com.example.grapefield.events.chat.model.request;
+package com.example.grapefield.chat.model.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
