+++ src/main/java/com/example/grapefield/events/model/entity/Events.java
@@ -1,6 +1,6 @@
 package com.example.grapefield.events.model.entity;
 
-import com.example.grapefield.events.chat.model.entity.ChatRoom;
+import com.example.grapefield.chat.model.entity.ChatRoom;
 import com.example.grapefield.events.participant.model.entity.EventsCast;
 import com.example.grapefield.events.participant.model.entity.EventsParticipation;
 import com.example.grapefield.events.post.model.entity.Board;
